% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoplot.conflibert_classify}
\alias{autoplot.conflibert_classify}
\title{Plot Classification Results with ggplot2}
\usage{
\method{autoplot}{conflibert_classify}(object, ...)
}
\arguments{
\item{object}{A result from \code{\link{conflibert_classify}}.}

\item{...}{Ignored.}
}
\value{
A ggplot object.
}
\description{
Aggregated view of a \code{\link{conflibert_classify}} result: how
many texts landed in each class, annotated with the average
confidence.
}
