% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finetune.R
\name{conflibert_models}
\alias{conflibert_models}
\title{List Available Models}
\usage{
conflibert_models()
}
\value{
A character vector of model names.
}
\description{
Returns the names of base models that can be used for fine-tuning
and comparison.
}
\examples{
conflibert_models()
}
