% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.conflibert_ner}
\alias{plot.conflibert_ner}
\title{Plot NER Results}
\usage{
\method{plot}{conflibert_ner}(x, ...)
}
\arguments{
\item{x}{A \code{conflibert_ner} tibble.}

\item{...}{Ignored.}
}
\value{
The object, invisibly.
}
\description{
Base-graphics aggregated view of a \code{\link{conflibert_ner}}
result: entity counts per type. For a ggplot2 version (including a
top-entities view) use \code{ggplot2::autoplot()}.
}
